# api/docs/swagger.yaml
basePath: /api
definitions:
...
  middlewares.Unauthorized:
    properties:
      error:
        type: string
    type: object
  models.DailyTraffic:
    properties:
      date:
        description: 'Format: YYYY-MM-DD'
        type: string
      rx:
        description: in GiB
        type: number
      tx:
        description: in GiB
        type: number
    type: object
...
  models.OcservUser:
    properties:
      config:
        $ref: '#/definitions/models.OcservUserConfig'
      created_at:
        type: string
      deactivated_at:
        type: string
      description:
        type: string
      expire_at:
        type: string
      group:
        type: string
      is_locked:
        type: boolean
      is_online:
        type: boolean
      password:
        type: string
      rx:
        description: Receive in bytes
        type: integer
      traffic_size:
        description: in GiB  >> x * 1024 ** 3
        type: integer
      traffic_type:
        enum:
        - Free
        - MonthlyTransmit
        - MonthlyReceive
        - TotallyTransmit
        - TotallyReceive
        type: string
      tx:
        description: Transmit in bytes
        type: integer
      uid:
        type: string
      updated_at:
        type: string
      username:
        type: string
    required:
    - created_at
    - group
    - is_locked
    - is_online
    - password
    - rx
    - traffic_size
    - traffic_type
    - tx
    - uid
    - username
    type: object
  models.OcservUserConfig:
    properties:
      banner:
        description: 'Text message shown to users when they connect to the VPN. Example:
          ''Welcome to the company VPN!'''
        type: string
      dns:
        description: 'Comma-separated list of DNS servers to assign to the user. Example:
          ''8.8.8.8,1.1.1.1'''
        items:
          type: string
        type: array
      explicit-ipv4:
        description: 'Static IPv4 address to assign to the user. Example: ''192.168.100.10'''
        type: string
      idle-timeout:
        description: 'Time in seconds before disconnecting idle users. Example: 600'
        type: integer
      ipv4-network:
        description: 'The pool of addresses from which to assign to the user. Example:
          ''192.168.1.0/24'''
        type: string
      iroute:
        description: 'Internal route available only via VPN. Example: ''10.0.0.0/8'''
        type: string
      mobile-idle-timeout:
        description: 'Idle timeout in seconds for mobile users. Example: 900'
        type: integer
      nbns:
        description: 'NetBIOS Name Servers (WINS) for Windows clients. Example: ''192.168.1.1'''
        type: string
      no-route:
        description: 'List of networks to exclude from VPN routing. Example: [''192.168.0.0/16'',
          ''10.0.0.0/8'']'
        items:
          type: string
        type: array
      rekey-time:
        description: 'Rekey time in seconds; triggers key renegotiation. Example:
          86400 for 24 hours'
        type: integer
      restrict-to-ports:
        description: 'Comma-separated list of allowed or blocked ports/protocols.
          Supports ''tcp(port)'', ''udp(port)'', ''icmp()'', ''icmpv6()'', and negation
          with ''!()''. Example: ''tcp(443), udp(53)'' or ''!(tcp(22), udp(1194))'''
        type: string
      restrict-to-routes:
        description: 'Allow user access only to defined routes. Example: true'
        type: boolean
      route:
        description: 'Routes pushed to the user for routing traffic. Example: [''0.0.0.0/0'',
          ''10.10.0.0/16'']'
        items:
          type: string
        type: array
      session-timeout:
        description: 'Maximum session time in seconds before forced disconnect. Example:
          3600'
        type: integer
      split-dns:
        description: 'List of domains over which the provided DNS servers should be
          used. Example: [''example.com'', ''internal.company.com'']'
        items:
          type: string
        type: array
    type: object
...
  ocserv_user.CreateOcservUserData:
    properties:
      config:
        $ref: '#/definitions/models.OcservUserConfig'
      description:
        example: User for testing VPN access
        maxLength: 1024
        type: string
      expire_at:
        example: "2025-12-31"
        type: string
      group:
        type: string
      password:
        maxLength: 32
        minLength: 2
        type: string
      traffic_size:
        description: 10 GiB
        example: 10737418240
        minimum: 0
        type: integer
      traffic_type:
        enum:
        - Free
        - MonthlyTransmit
        - MonthlyReceive
        - TotallyTransmit
        - TotallyReceive
        example: MonthlyTransmit
        type: string
      username:
        maxLength: 32
        minLength: 2
        type: string
    required:
    - group
    - password
    - traffic_type
    - username
    type: object
  ocserv_user.OcservUsersResponse:
    properties:
      meta:
        $ref: '#/definitions/request.Meta'
      result:
        items:
          $ref: '#/definitions/models.OcservUser'
        type: array
    required:
    - meta
    type: object
  ocserv_user.UpdateOcservUserData:
    properties:
      config:
        $ref: '#/definitions/models.OcservUserConfig'
      description:
        example: User for testing VPN access
        maxLength: 1024
        type: string
      expire_at:
        example: "2025-12-31"
        type: string
      group:
        example: default
        type: string
      password:
        example: strongpassword123
        maxLength: 32
        minLength: 2
        type: string
      traffic_size:
        description: 10 GiB
        example: 10737418240
        type: integer
      traffic_type:
        enum:
        - Free
        - MonthlyTransmit
        - MonthlyReceive
        - TotallyTransmit
        - TotallyReceive
        example: MonthlyTransmit
        type: string
    type: object
  request.ErrorResponse:
    properties:
      error:
        items:
          type: string
        type: array
      message:
        items:
          type: string
        type: array
    required:
    - error
    - message
    type: object
  request.Meta:
    properties:
      page:
        type: integer
      size:
        type: integer
      total_records:
        type: integer
    required:
    - page
    - size
    - total_records
    type: object
...
info:
  contact: {}
  description: This is a sample Ocserv User management Api server.
  title: Ocserv User management Example Api
  version: "1.0"
paths:
...
  /ocserv/users:
    get:
      consumes:
      - application/json
      description: List of Ocserv Users
      parameters:
      - description: Page number, starting from 1
        in: query
        minimum: 1
        name: page
        type: integer
      - description: Number of items per page
        in: query
        maximum: 100
        minimum: 1
        name: size
        type: integer
      - description: Field to order by
        in: query
        name: order
        type: string
      - description: Sort order, either ASC or DESC
        enum:
        - ASC
        - DESC
        in: query
        name: sort
        type: string
      - description: Bearer TOKEN
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ocserv_user.OcservUsersResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/request.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/middlewares.Unauthorized'
      summary: List of Ocserv Users
      tags:
      - Ocserv(Users)
    post:
      consumes:
      - application/json
      description: Ocserv User creation
      parameters:
      - description: Bearer TOKEN
        in: header
        name: Authorization
        required: true
        type: string
      - description: ocserv user create data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/ocserv_user.CreateOcservUserData'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.OcservUser'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/request.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/middlewares.Unauthorized'
      summary: Ocserv User creation
      tags:
      - Ocserv(Users)
  /ocserv/users/{uid}:
    delete:
      consumes:
      - application/json
      description: Ocserv User delete
      parameters:
      - description: Bearer TOKEN
        in: header
        name: Authorization
        required: true
        type: string
      - description: Ocserv User UID
        in: path
        name: uid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/request.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/middlewares.Unauthorized'
      summary: Ocserv User delete
      tags:
      - Ocserv(Users)
    patch:
      consumes:
      - application/json
      description: Ocserv User update
      parameters:
      - description: Bearer TOKEN
        in: header
        name: Authorization
        required: true
        type: string
      - description: Ocserv User UID
        in: path
        name: uid
        required: true
        type: string
      - description: ocserv user update data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/ocserv_user.UpdateOcservUserData'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.OcservUser'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/request.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/middlewares.Unauthorized'
      summary: Ocserv User update
      tags:
      - Ocserv(Users)
  /ocserv/users/{uid}/lock:
    post:
      consumes:
      - application/json
      description: Ocserv User locking
      parameters:
      - description: Bearer TOKEN
        in: header
        name: Authorization
        required: true
        type: string
      - description: Ocserv User UID
        in: path
        name: uid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/request.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/middlewares.Unauthorized'
      summary: Ocserv User locking
      tags:
      - Ocserv(Users)
  /ocserv/users/{uid}/statistics:
    get:
      consumes:
      - application/json
      description: Ocserv User Statistics
      parameters:
      - description: Bearer TOKEN
        in: header
        name: Authorization
        required: true
        type: string
      - description: Ocserv User UID
        in: path
        name: uid
        required: true
        type: string
      - description: date_start
        in: query
        name: date_start
        type: string
      - description: date_end
        in: query
        name: date_end
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.DailyTraffic'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/request.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/middlewares.Unauthorized'
      summary: Ocserv User Statistics
      tags:
      - Ocserv(Users)
  /ocserv/users/{uid}/unlock:
    post:
      consumes:
      - application/json
      description: Ocserv User unlocking
      parameters:
      - description: Bearer TOKEN
        in: header
        name: Authorization
        required: true
        type: string
      - description: Ocserv User UID
        in: path
        name: uid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/request.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/middlewares.Unauthorized'
      summary: Ocserv User unlocking
      tags:
      - Ocserv(Users)
  /ocserv/users/{username}/disconnect:
    post:
      consumes:
      - application/json
      description: Disconnect Ocserv User
      parameters:
      - description: Bearer TOKEN
        in: header
        name: Authorization
        required: true
        type: string
      - description: Ocserv User username
        in: path
        name: username
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/request.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/middlewares.Unauthorized'
      summary: Disconnect Ocserv User
      tags:
      - Ocserv(Users)
  /ocserv/users/statistics:
    get:
      consumes:
      - application/json
      description: Ocserv Users Statistics
      parameters:
      - description: Bearer TOKEN
        in: header
        name: Authorization
        required: true
        type: string
      - description: date_start
        in: query
        name: date_start
        required: true
        type: string
      - description: date_end
        in: query
        name: date_end
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.DailyTraffic'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/request.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/middlewares.Unauthorized'
      summary: Ocserv Users Statistics
      tags:
      - Ocserv(Statistics)
...
swagger: "2.0"